requirements.txt
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.0
bleach==5.0.1
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-ipware==4.0.2
django-model-utils==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-api-logger==1.1.10
drf-spectacular==0.24.0
et-xmlfile==1.1.0
inflection==0.5.1
jsonschema==4.16.0
mysqlclient==2.1.1
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
webencodings==0.5.1
